refactor(mitm-addon): extract request classification owner
